requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-generator==1.10
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
Django==4.1.7
django-cors-headers==3.14.0
django-crontab==0.7.1
django-dbbackup==4.0.2
django-extensions==3.2.3
djangorestframework==3.14.0
docopt==0.6.2
exceptiongroup==1.1.1
h11==0.14.0
html5lib==1.1
idna==3.4
mysql-connector-python==8.3.0
num2words==0.5.12
oscrypto==1.3.0
outcome==1.2.0
packaging==23.1
pdfcrowd==5.13.0
Pillow==9.5.0
pycparser==2.21
pydot==1.4.2
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pyhtml2pdf==0.0.6
PyJWT==2.6.0
PyParse==1.1.7
pyparsing==3.1.0
pypdf==3.9.0
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.2
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.2
tzdata==2022.7
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.15
webdriver-manager==3.8.6
webencodings==0.5.1
whitenoise==6.4.0
wsproto==1.2.0
xhtml2pdf==0.2.11
